Moved OneDim code to the test module instead
